+++ HARK/distribution.py
@@ -8,6 +8,8 @@ import scipy.stats as stats
 
 class Distribution:
     """
+    Base class for all probability distributions.
+
     Parameters
     ----------
     seed : int
@@ -27,6 +29,147 @@ class Distribution:
         """
         self.RNG = np.random.RandomState(self.seed)
 
+class IndexDistribution(Distribution):
+    """
+    This class provides a way to define a distribution that
+    is conditional on an index.
+
+    The current implementation combines a defined distribution
+    class (such as Bernoulli, LogNormal, etc.) with information
+    about the conditions on the parameters of the distribution.
+
+    For example, an IndexDistribution can be defined as
+    a Bernoulli distribution whose parameter p is a function of
+    a different inpute parameter.
+
+    Parameters
+    ----------
+
+    engine : Distribution class
+        A Distribution subclass.
+
+    conditional: dict
+        Information about the conditional variation
+        on the input parameters of the engine distribution.
+        Keys should match the arguments to the engine class
+        constructor.
+
+    seed : int
+        Seed for random number generator.
+    """
+    conditional = None
+    engine = None
+
+    def __init__(self, engine, conditional, seed = 0):
+        # Set up the RNG
+        super().__init__(seed)
+
+        self.conditional = conditional
+        self.engine = engine
+
+    def __getitem__(self, y):
+        # test one item to determine case handling
+        item0 = list(self.conditional.values())[0]
+
+        if type(item0) is list:
+            cond = {key : val[y] for (key, val) in self.conditional.items()}
+            return self.engine(
+                    seed=self.RNG.randint(0, 2 ** 31 - 1),
+                    **cond
+                )
+        else:
+            raise(Exception(
+                f"IndexDistribution: Unhandled case for __getitem__ access. y: {y}; conditional: {conditional}"
+                ))
+
+    def approx(self, N, **kwds):
+        """
+        Approximation of the distribution.
+
+        Parameters
+        ----------
+        N : init
+            Number of discrete points to approximate
+            continuous distribution into.
+
+        kwds: dict
+            Other keyword arguments passed to engine
+            distribution approx() method.
+
+        Returns:
+        ------------
+        dists : [DiscreteDistribution]
+            A list of DiscreteDistributions that are the
+            approximation of engine distribution under each condition.
+
+            TODO: It would be better if there were a conditional discrete
+            distribution representation. But that integrates with the
+            solution code. This implementation will return the list of
+            distributions representations expected by the solution code.
+        """
+
+        # test one item to determine case handling
+        item0 = list(self.conditional.values())[0]
+
+        if type(item0) is float:
+            # degenerate case. Treat the parameterization as constant.
+            return self.engine(
+                seed = self.RNG.randint(0, 2 ** 31 - 1),
+                **self.conditional
+                ).approx(N,**kwds)
+
+        if type(item0) is list:
+            return [self[i].approx(N,**kwds) for i, _ in enumerate(item0)]
+
+
+    def draw(self, condition):
+        """
+        Generate arrays of draws.
+        The input is an array containing the conditions.
+        The output is an array of the same length (axis 1 dimension)
+        as the conditions containing random draws of the conditional
+        distribution.
+
+        Parameters
+        ----------
+        condition : np.array
+            The input conditions to the distribution.
+
+        Returns:
+        ------------
+        draws : np.array
+        """
+        # for now, assume that all the conditionals
+        # are of the same type.
+        # this matches the HARK 'time-varying' model architecture.
+
+        # test one item to determine case handling
+        item0 = list(self.conditional.values())[0]
+
+        if type(item0) is float:
+            # degenerate case. Treat the parameterization as constant.
+            N = condition.size
+
+            return self.engine(
+                seed = self.RNG.randint(0, 2 ** 31 - 1),
+                **self.conditional
+                ).draw(N)
+
+        if type(item0) is list:
+            # conditions are indices into list
+            # somewhat convoluted sampling strategy retained
+            # for test backwards compatibility
+
+            draws = np.zeros(condition.size)
+
+            for c in np.unique(condition):
+                these = c == condition
+                N = np.sum(these)
+
+                cond = {key : val[c] for (key, val) in self.conditional.items()}
+                draws[these] = self[c].draw(N)
+
+            return draws
 
 ### CONTINUOUS DISTRIBUTIONS
 
@@ -38,11 +181,11 @@ class Lognormal(Distribution):
     Parameters
     ----------
     mu : float or [float]
-        One or more means.  Number of elements T in mu determines number
-        of rows of output.
+        One or more means of underlying normal distribution.
+        Number of elements T in mu determines number of rows of output.
     sigma : float or [float]
-        One or more standard deviations. Number of elements T in sigma
-        determines number of rows of output.
+        One or more standard deviations of underlying normal distribution.
+        Number of elements T in sigma determines number of rows of output.
     seed : int
         Seed for random number generator.
     """
@@ -64,7 +207,7 @@ class Lognormal(Distribution):
 
     def draw(self, N):
         """
-        Generate arrays of lognormal draws. The sigma input can be a number
+        Generate arrays of lognormal draws. The sigma parameter can be a number
         or list-like.  If a number, output is a length N array of draws from the
         lognormal distribution with standard deviation sigma. If a list, output is
         a length T list whose t-th entry is a length N array of draws from the
@@ -74,8 +217,6 @@ class Lognormal(Distribution):
         ----------
         N : int
             Number of draws in each row.
-        seed : int
-            Seed for random number generator.
 
         Returns:
         ------------